Allows for clicking on nested DOM elements within each .text-suggestion
When using a custom `render()` method containing more than 1 level of
nesting, the `onClick` event is not triggered when not clicking
*directly* on the root `.text-suggestion` element.

 - http://goo.gl/MpY8f

Clicking anywhere in the green box will not trigger the `onClick`.
Clicking in the red box *(a small sliver at the bottom of the
`.text-suggestion`)* **does** trigger the `onClick`.

This changeset causes the `onMouseOver` and `onClick` events to look at
the `target`'s ancestor's to see if one of them is a `.text-suggestion`,
swapping the `target` to that DOM element if found.